refactor: 메소드 정리
